requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
googleapis-common-protos==1.59.0
idna==3.4
importlib-metadata==6.0.1
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
pkg_resources==0.0.0
protobuf==4.22.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
